requirements.txt
streamlit
openai
groq
pinecone
python-dotenv
sentence_transformers